requirements.txt
django==5.0.1
celery==5.3.6
# django-celery-beat==2.5.0 https://github.com/celery/django-celery-beat/issues/680
django-environ==0.11.2
django-admin-site-search==0.4.0

# https://docs.djangoproject.com/en/4.0/topics/cache/#redis
redis==5.0.1
hiredis==2.3.2

psycopg[binary]==3.1.17

gunicorn==21.2.0
uvicorn==0.27.0
whitenoise==6.6.0

sentry-sdk==1.39.2

# dev dependencies
django-debug-toolbar==4.2.0
django-extensions==3.2.3
factory-boy==3.3.0
pytest==7.4.4
pytest-django==4.7.0
ipython==8.20.0
